WIP - refactor and cleanup
